+++ compiler/src/main/kotlin/dev/zacsweers/metro/compiler/fir/checkers/BindingContainerCallableChecker.kt
@@ -13,7 +13,7 @@ import dev.zacsweers.metro.compiler.fir.MetroDiagnostics.BINDING_CONTAINER_ERROR
 import dev.zacsweers.metro.compiler.fir.annotationsIn
 import dev.zacsweers.metro.compiler.fir.classIds
 import dev.zacsweers.metro.compiler.fir.compatContext
-import dev.zacsweers.metro.compiler.fir.findInjectLikeConstructors
+import dev.zacsweers.metro.compiler.fir.findInjectConstructors
 import dev.zacsweers.metro.compiler.fir.isAnnotatedWithAny
 import dev.zacsweers.metro.compiler.fir.metroFirBuiltIns
 import dev.zacsweers.metro.compiler.fir.scopeAnnotations
@@ -336,8 +336,7 @@ internal object BindingContainerCallableChecker :
 
       if (returnType.typeArguments.isEmpty()) {
         val returnClass = returnType.toClassSymbol(session) ?: return
-        // TODO after we migrate fully to AssistedInject, allow qualified assisted inject types
-        val injectConstructor = returnClass.findInjectLikeConstructors(session).firstOrNull()
+        val injectConstructor = returnClass.findInjectConstructors(session).firstOrNull()
 
         if (injectConstructor != null) {
           // If the type keys and scope are the same, this is redundant