+++ gradle-plugin/src/main/kotlin/dev/zacsweers/metro/gradle/MetroGradleSubplugin.kt
@@ -166,12 +166,6 @@ public class MetroGradleSubplugin : KotlinCompilerPluginSupportPlugin {
         add(lazyOption("public-provider-severity", extension.publicProviderSeverity))
         add(
           lazyOption(
-            "assisted-inject-deprecation-severity",
-            extension.assistedInjectMigrationSeverity,
-          )
-        )
-        add(
-          lazyOption(
             "warn-on-inject-annotation-placement",
             extension.warnOnInjectAnnotationPlacement,
           )